requirements.txt
azureml-sdk==1.0.30
jupyter
environs==4.1.0
mlflow==0.9.1
numpy==1.16.2
pandas==0.24.1
python-dotenv==0.10.1
scipy==1.2.1
scikit-learn==0.20.1
azure-storage